CBSI-160: Added HAB support for kernel
A HAB kernel image is a zImage with a 32 byte IVT + a CSF appended to
it. Due to padding requirement for the image (4k) and variable length
CSF, it becomes tricky to determine the start of the IVT header only
using the file size.
Solving this by exporting the IVT offset directly using a uboot control
file genreated by the signing tool.