CBSI-160: Check HAB fuse and adjust environment
If HAB is fused, we need to set the zimage variable and bootscript
accordingly. By doing this runtime we are able to support non-secure
images on a unblown HAB fuse boards and at the same time secure images.

In other words: Do not break anything already out in the wild.